Started this refactoring session by looking at evaluate_elements. It's a bit too big. It seems to be reponible for looping and summing the elements and filtering whether we should be adding this element. There's also a bit of mess around the use of the sum variable to store and return the value computed in the string.
The first thing that strikes me in the Ruby toolbox is that although I'm
using a sum variable initialized outside a loop and I'm reassigning it
inside the loop, this is the kind of thing that Ruby's inject method
excels at. By introducing this, I can clear up the use of the sum
variable so that I've got fewer syntactic moving pieces in this method.

I run the specs again, but now find that
StringCalculator calc ignores numbers greater than 1000
isn't passing any longer. We're getting a nil back instead of a 0