Just looking over the evaluate_elements with the intention of getting it to read more like a declaration of intent, a couple of the most jarring things are the || 0 and inject(0).
These had been put in place at the time when evaluate_elements had the
responsibility to sum numbers encoded as strings which might evaluate to
nil. Now that we've moved the responsibilty of collecting the valid
numbers from the input, we can guarantee that anything passed to inject
will be a valid number. That means that anything coming out of the
inject will also be a valid number. The result is that we can remove
these redundant safety checks and improve the readability of the code.