Looking at the StringCalculator, a lot of the methods aren't to do with calculation in strings as such. Instead, they're to do with reading the cutom delimiters from the string and then returning a list of tokens for the rest of the StringCalculator to then add together, throw an exception over or ignore completely.
It looks like this class is doing too much, so I've introduced a
Tokenizer. As I did so, I realized that it was a bit of a specific
Tokenizer in that it uses an uncommon convention for specifying
delimiters as part of the token input itself. Since I don't want to risk
someone picking this up, I've introduced a StringCalculator module to
wrap up both the StringCalculator and Tokenizer classes.

I've started by moving the functionality and tests associated with
returning tokens from the input into their own class and spec. I then
refactored the StringCalculator to use my new Tokenizer.

At this stage StringCalculator is really a test that integrates the
functionality of both the StringCalculator and the Tokenizer.

All the tests passed, so we're good to continue.